feat: apply `convert_env` to remaining endpoints
Let's use this url as example:

http://domain.env.com/api/endpoint

There are 3 different environments in MZ:
1. production
2. staging
3. integration/UAT (as EDG calls it)

However, only production and staging are mapped to the url, being converted to `prod` and `stage` respectivelly, so integration (`integ`) must be converted before the request is sent, otherwise, it will return an error.

`convert_env` is a function to check if `env` equals to `integ` and, if so, converts it to `prod`. In any case, `convert_env` also guarantees only lowercase environments are used in URLs.